Update simbrief_form.blade.php
Found a flaw in the logic resulting 0 pax info sent to simbrief with a load distribution text, which causes simbrief to calculate a flight with no pax.

Issue was occuring if only the user has no fare settings done and/or assigned to the subfleet.

In such cases simbrief will generate a random load for the flight plan calculation, we will not be sending any data to block it's logic.