Add atc attribute to user model
Same as flight model, if a user has an alphanumeric callsign, VA's should be able to use it easily too when needed.

If no callsign is defined, ident is used.

(padding with zero's are not used in atc attribute 'cause this can be just FDX1 or DLH72 without a leading zero as atc callsign)