Update FlightService.php
A flight should NOT be marked as duplicate unless

* Airline
* Flight Number
* Route Code
* Route Leg Number
* Origin
* Destination
* Scheduled Days

are identical with another flight.

We can import flights with same numbers but with different origins and destinations, but can not edit them in admin due to very limited "duplicate check".

Imagine you have a flight

TK1 AYT-DUS
TK1 DUS-BER
TK2 BER-AYT

You can import this, but can not edit TK1 because v7 thinks it has a duplicate.

Change extends the check properly to consider origin, destination and weekdays.